private_core/requirements.txt
# TradeOracle Nexus - Python Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Elasticsearch
elasticsearch==8.17.0

# AI / LLM
langchain==0.3.14
langchain-elasticsearch==0.3.0
langchain-community==0.3.14

# HTTP & Data
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4

# Trading Data
ccxt==4.4.50

# Telegram
python-telegram-bot==21.9

# Utilities
rich==13.9.4
